added deletion, creation in DatabaseManager

+++ src/main/java/com/netcreker/meetup/entitymanager/EntityManager.java
@@ -1,7 +1,7 @@
 package com.netcreker.meetup.entitymanager;
 
 public interface EntityManager {
-    <T> T load(int id, Class<T> clazz);
-    void delete(int id);
+    <T> T load(long id, Class<T> clazz);
+    void delete(long id);
     void save(Object instance, Class<?> clazz);
 }
